+++ R/pxw_data.R
@@ -16,7 +16,7 @@
 #' @param to_name Whether to add columns for names. TRUE, FALSE (default), or a vector of variable names.
 #' @param name_suffix Suffix to add to name columns. Default is `"_name"`.
 #' @param code_suffix Suffix to add to code columns. Default is `""`.
-#' @param time_format A time format for \link{parse_times}.
+#' @param date_format A time format for \link{parse_dates}.
 #'        See also \link[lubridate]{parse_date_time}.
 #'
 #' @import pxweb
@@ -36,7 +36,7 @@
 #' data_named <- pxw_get_data(url, query, to_name = TRUE)
 pxw_get_data <- function(url, query, to_name = FALSE,
                          name_suffix = "_name", code_suffix = "",
-                         time_format = NULL) {
+                         date_format = NULL) {
 
   # Fetch data from PxWeb API
   px_data <- pxweb::pxweb_get(url = url, query = query)
@@ -60,7 +60,7 @@ pxw_get_data <- function(url, query, to_name = FALSE,
     tidyr::pivot_longer(where(is.numeric),
                         names_to = setdiff(names(codes_names), names(.)),
                         values_to = "values") %>%
-    parse_times(time_format = time_format) %>%
+    parse_dates(date_format = date_format) %>%
     statfitools::codes2names(codes_names, to_name = to_name,
                              name_suffix = name_suffix,
                              code_suffix = code_suffix) %>%